requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
env-file==2020.5.9
pkg-resources==0.0.0
public==2019.4.13
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
values==2019.4.13
